pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "ananas-ai"
version = "0.1.0"
description = "Ananas AI Platform — marketing intelligence system"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.84.0",
    "python-dotenv>=1.0.0",
    "google-analytics-data>=0.20.0",
    "watchdog>=6.0.0",
    "pypdf>=5.0.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
    "pillow>=11.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
ananas-ai = "ananas_ai.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

# ── Pytest ────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths  = ["tests"]
addopts    = "-v --tb=short --cov=ananas_ai --cov-report=term-missing --cov-report=xml"

# ── Ruff — linter + formatter (replaces flake8 + black + isort) ───────────────
[tool.ruff]
target-version = "py310"
line-length    = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long — ruff format handles this
    "B008",  # function call in default arg — common pattern here
]

[tool.ruff.lint.isort]
known-first-party = ["ananas_ai"]

# ── Mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version         = "3.10"
warn_return_any        = true
warn_unused_configs    = true
ignore_missing_imports = true   # external stubs not always available
